Migrated ab indicator to class pattern.
