+++ js/masters/indicators/acceleration-bands.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/ABIndicator.js';
+import '../../Stock/Indicators/AB/ABIndicator.js';
