+++ ts/Stock/Indicators/AB/ABPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ABandsIndicator from './ABIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class ABandsPoint extends SMAPoint {
+    public series: ABandsIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default ABandsPoint;
